Completado de funciones en service. En persistance, ahora devuelven sus respectivos objetos

+++ WizardUniversityOfNikolaus/Service/ProfessorService.cs
@@ -5,24 +5,51 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 
 namespace Service
 {
     public class ProfessorService 
     {
-        public async Task CreateAsync(Professor professor)//TODO le tengo que pasar los datos no un profesor
+        private bool InputCheck(string name, int age)
         {
-            await DataBase.GetInstance().professorRepository.Create(professor);
+            if (name == "")
+            {
+                throw new Exception("You must complete the name space");
+            }
+            if (age < 18)
+            {
+                throw new Exception("The age number must be more than 18. If not, it is child labour! Call 911");
+            }
+            return true;
         }
 
-        public Task<bool> DeleteAsync(int id)
+        public async Task<Professor> CreateAsync(string name, int age)//TODO le tengo que pasar los datos no un profesor
         {
-            return DataBase.GetInstance().professorRepository.DeleteAsync(id);
+            if(!InputCheck(name, age)) { throw new Exception(); }
+            Professor professor = new Professor(name, age);
+            return await DataBase.GetInstance().professorRepository.CreateAsync(professor);
         }
 
-        public Task<bool> UpdateAsync(Professor professor)//TODO Aca también los datos
+        public async Task<bool> DeleteAsync(int id, string name, int age)
         {
-            return DataBase.GetInstance().professorRepository.UpdateAsync(professor);
+            if (!await DataBase.GetInstance().professorRepository.ValidateInfoAsync(id, name, age))
+            {
+                throw new Exception("The input is wrong, please check the info provided");
+            }
+            return await DataBase.GetInstance().professorRepository.DeleteAsync(id);
+        }
+
+        public async Task<Professor> UpdateAsync(int idProfessor, string oldName, int oldAge, string newName, int newAge)//TODO Aca también los datos
+        {
+            if (!await DataBase.GetInstance().professorRepository.ValidateInfoAsync(idProfessor, oldName, oldAge))
+            {
+                throw new Exception("The input is wrong, please check the info provided");
+            }
+            if(!InputCheck(newName, newAge)) { throw new Exception(); }
+            Professor professor = new Professor(newName, newAge, idProfessor );
+            return await DataBase.GetInstance().professorRepository.UpdateAsync(professor); //Cambia las caracteristicas de professor?
+            
         }
 
         public async Task GetStudentProfessorsAsync(int idStudent)
